Made an equality operator on Axis2D (and by interdependence, on Histo2D) to work in a bit more intelligent way, if the Axis2D describes a whole grid (it checks if there exists a bin in the Axis 2D on the right side of an equality sign at a location being considered that has the same dimensions as a bin on the left side of equality sign and if it is true for all the bins then return that two Histos/Axises are the same). For a more reasonable explanation, please see the code in Axis2D.h, ~line 485. It is not yet tested, as noted in the comments.
I really, really don't want to do it in the same way when an axis is not a full grid, because in such case we need to do O(n^2) comparisons which with 40k bins (200x200) would take forever. Or I think it would, it is yet to be tested.